Beginning work on Timers
This will be extended into a reconstruction of UI classes and transitions. The plan is to allow every UI elements a potential to implement whatever kind of tick they want. I also want to make more explicit classes. This should be a better way to implement picking as well, which is what this leads into.